Minor test improvements
